+++ internal/logic/indexer/indexer_abelian.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/b2network/b2-indexer/internal/config"
+	"github.com/b2network/b2-indexer/config"
 	"github.com/b2network/b2-indexer/internal/types"
 	"github.com/b2network/b2-indexer/pkg/log"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
