+++ internal/logic/indexer/indexer_abelian_test.go
@@ -3,7 +3,7 @@ package indexer
 import (
 	"testing"
 
-	"github.com/b2network/b2-indexer/internal/config"
+	"github.com/b2network/b2-indexer/config"
 	logger "github.com/b2network/b2-indexer/pkg/log"
 )
