+++ internal/logic/rollup/rollup_service.go
@@ -6,12 +6,12 @@ import (
 	"math/big"
 	"time"
 
+	"github.com/b2network/b2-indexer/config"
 	"github.com/b2network/b2-indexer/pkg/event"
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/shopspring/decimal"
 
-	"github.com/b2network/b2-indexer/internal/config"
 	"github.com/b2network/b2-indexer/internal/model"
 	"github.com/b2network/b2-indexer/pkg/log"
 	"github.com/cometbft/cometbft/libs/service"
