+++ internal/model/deposit.go
@@ -26,19 +26,6 @@ const (
 )
 
 const (
-	DepositB2EoaTxStatusSuccess         = iota // eoa transfer success
-	DepositB2EoaTxStatusPending                // eoa transfer pending
-	DepositB2EoaTxStatusFailed                 // eoa transfer failed
-	DepositB2EoaTxStatusWaitMinedFailed        // eoa transfer wait mined failed
-	DepositB2EoaTxStatusWaitMined
-	_
-	_
-	DepositB2EoaTxStatusContextDeadlineExceeded // eoa transfer client context deadline exceeded
-	DepositB2EoaTxStatusUnknown
-	DepositB2EoaTxStatusNonceToLow
-)
-
-const (
 	CallbackStatusSuccess = iota
 	CallbackStatusPending
 )
@@ -71,10 +58,6 @@ type Deposit struct {
 	B2TxNonce        uint64    `json:"b2_tx_nonce" gorm:"default:0"`
 	B2TxStatus       int       `json:"b2_tx_status" gorm:"type:SMALLINT;default:1"`
 	B2TxRetry        int       `json:"b2_tx_retry" gorm:"type:SMALLINT;default:0"`
-	B2EoaTxFrom      string    `json:"b2_eoa_tx_from" gorm:"type:text;default:'';comment:from address"`
-	B2EoaTxNonce     uint64    `json:"b2_eoa_tx_nonce" gorm:"default:0"`
-	B2EoaTxHash      string    `json:"b2_eoa_tx_hash" gorm:"type:text;not null;default:'';comment:b2 network eoa tx hash"`
-	B2EoaTxStatus    int       `json:"b2_eoa_tx_status" gorm:"type:SMALLINT;default:1"`
 	BtcBlockTime     time.Time `json:"btc_block_time"`
 	CallbackStatus   int       `json:"callback_status" gorm:"type:SMALLINT;default:0"`
 	ListenerStatus   int       `json:"listener_status" gorm:"type:SMALLINT;default:0"`
@@ -97,10 +80,6 @@ type DepositColumns struct {
 	B2TxNonce        string
 	B2TxStatus       string
 	B2TxRetry        string
-	B2EoaTxFrom      string
-	B2EoaTxNonce     string
-	B2EoaTxHash      string
-	B2EoaTxStatus    string
 	BtcBlockTime     string
 	CallbackStatus   string
 	ListenerStatus   string
@@ -127,10 +106,6 @@ func (Deposit) Column() DepositColumns {
 		B2TxHash:         "b2_tx_hash",
 		B2TxNonce:        "b2_tx_nonce",
 		B2TxStatus:       "b2_tx_status",
-		B2EoaTxFrom:      "b2_eoa_tx_from",
-		B2EoaTxNonce:     "b2_eoa_tx_nonce",
-		B2EoaTxHash:      "b2_eoa_tx_hash",
-		B2EoaTxStatus:    "b2_eoa_tx_status",
 		BtcBlockTime:     "btc_block_time",
 		B2TxRetry:        "b2_tx_retry",
 		CallbackStatus:   "callback_status",
