+++ main.go
@@ -3,7 +3,9 @@ Copyright © 2024 NAME HERE <EMAIL ADDRESS>
 */
 package main
 
-import "github.com/b2network/b2-indexer/internal/cmd"
+import (
+	"github.com/b2network/b2-indexer/cmd"
+)
 
 func main() {
 	cmd.Execute()
